Resolve stylecop warnings on Bot.Builder.AI.LUIS.Tests
