+++ tests/Microsoft.Bot.Builder.AI.LUIS.Tests/TelemetryOverrideRecognizer.cs
@@ -1,0 +1,54 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net.Http;
+using System.Reflection;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Bot.Builder.Adapters;
+using Microsoft.Bot.Configuration;
+using Microsoft.Bot.Schema;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Moq;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using RichardSzalay.MockHttp;
+
+namespace Microsoft.Bot.Builder.AI.Luis.Tests
+{
+
+    public class TelemetryOverrideRecognizer : LuisRecognizer
+    {
+        public TelemetryOverrideRecognizer(IBotTelemetryClient telemetryClient, LuisApplication application, LuisPredictionOptions predictionOptions = null, bool includeApiResults = false, bool logPersonalInformation = false, HttpClientHandler clientHandler = null)
+           : base(application, predictionOptions, includeApiResults, clientHandler)
+        {
+            LogPersonalInformation = logPersonalInformation;
+        }
+
+        protected override Task OnRecognizerResultAsync(RecognizerResult recognizerResult, ITurnContext turnContext, Dictionary<string, string> properties = null, Dictionary<string, double> metrics = null, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            properties.TryAdd("MyImportantProperty", "myImportantValue");
+
+            // Log event
+            TelemetryClient.TrackEvent(
+                            LuisTelemetryConstants.LuisResult,
+                            properties,
+                            metrics);
+
+            // Create second event.
+            var secondEventProperties = new Dictionary<string, string>();
+            secondEventProperties.Add(
+                "MyImportantProperty2",
+                "myImportantValue2");
+            TelemetryClient.TrackEvent(
+                            "MySecondEvent",
+                            secondEventProperties);
+            return Task.CompletedTask;
+        }
+    }
+}
